+++ tools/tools.go
@@ -21,7 +21,6 @@ import (
 	_ "github.com/client9/misspell/cmd/misspell"
 	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
 	_ "github.com/jcchavezs/porto/cmd/porto"
-	_ "github.com/mjibson/esc"
 	_ "github.com/wadey/gocovmerge"
 	_ "go.opentelemetry.io/build-tools/crosslink"
 	_ "go.opentelemetry.io/build-tools/dbotconf"